More work done on Actions, It can now be edited, removed, save, loaded, all works. Need to implement error handling Settings system added, need to be used Logging system implemented, need to be saved as a file

+++ Tao Bot Maker/Model/SequenceRepository.cs
@@ -29,33 +29,45 @@ namespace Tao_Bot_Maker.Model
             return sequenceNames;
         }
 
-        public void RemoveSequence(string name)
+        public bool RemoveSequence(string name)
         {
+            if (name == null) return false;
+
             string filePath = Path.Combine(sequencesFolderPath, $"{name}.json");
             if (File.Exists(filePath))
             {
                 File.Delete(filePath);
+                return true;
             }
+            return false;
         }
 
         public Sequence GetSequence(string name)
         {
+            if (name == null) return null;
+
             string filePath = Path.Combine(sequencesFolderPath, $"{name}.json");
             if (File.Exists(filePath))
             {
-                string json = File.ReadAllText(filePath);
-                return JsonConvert.DeserializeObject<Sequence>(json, new JsonSerializerSettings
+                try
                 {
-                    TypeNameHandling = TypeNameHandling.Auto,
-                    ReferenceLoopHandling = ReferenceLoopHandling.Ignore
-                });
+                    string json = File.ReadAllText(filePath);
+                    return JsonConvert.DeserializeObject<Sequence>(json, new JsonSerializerSettings
+                    {
+                        TypeNameHandling = TypeNameHandling.Auto,
+                        ReferenceLoopHandling = ReferenceLoopHandling.Ignore
+                    });
+                } catch
+                {
+                    throw new Exception("Error loading sequence");
+                }
             }
             return null;
         }
 
         public void SaveSequence(Sequence sequence, string name)
         {
-            string filePath = Path.Combine(sequencesFolderPath, $"{name}.json"); 
+            string filePath = Path.Combine(sequencesFolderPath, $"{name}.json");
             string json = JsonConvert.SerializeObject(sequence, Formatting.Indented, new JsonSerializerSettings
             {
                 TypeNameHandling = TypeNameHandling.Auto,
